NEO: optimize the cNEO root finding procedure
When finding `f` values, only `f` changes in the nuclear core
Hamiltonian. However, the previous implementation re-evaluates all the
matrices (core, veff, etc) every time `scipy.optimize.root` enters the
derivative function, which is completely unnecessary and introduced much
computational overhead. Now optimize it by storing these matrices in the
first step and pass those variables to the function.

In the SCF2 version, it is also unnecessary to re-diagonalize the matrix
to retrieve the density matrix, as the density matrix was already
computed in the function. Change to directly use it, thus avoiding the
additional diagonalization.

In addition, verbosity flag initialization is also introduced, as what
had been done for HF and KS.