NEO: remove the redundant h_ep separation
Previously, it was believed that the separation between veff and h_ep
was necessary because veff does not need to update if only the densities
of other particles have changed. However, the fact is that we can
usually delay the update of veff until h_ep has to be updated, and h_ep
depends on the densities of both kinds of particles, not just one kind.
This commit reverts to the old behavior that veff contains h_ep. The
core Hamiltonian is the only exception that depends on the other
particles' densities only.

The convergence behavior is not entirely the same as before this commit,
as h_ep in the nuclear part is not updated immediately as the electronic
density changes (but on the other hand, it is properly updated when the
nuclear density changes, which was not done before). This makes the EPC
tests fail again, but this is likely due to some other reasons that make
EPC calculations hard.