NEO: proper DIIS setup
In the previous implementation, DIIS was incorrectly applied to
electronic and nuclear equations separately. This caused significant
oscillations when electron-proton correlation is enabled. In practice,
just as how unrestricted calculations are dealt with, all Fock matrices
should share the same set of extrapolation coefficients, and this set of
coefficients should be solved from the error vector that takes errors
from all Fock matrices into account.

To do this, the DIIS class was modified to support a list of Fock
matrices (as electronic and nuclear Fock matrices are usually of
different sizes). For now, only the original CDIIS is supported.

We notice that when no epc is used, the original scheme seems to work
(nuclear part has no DIIS), but we still advocate using this new scheme
as it improves the performance of SCF even when no epc is present.

Also, in this commit no CDFT is implemented, as it will be left to the
next commit.