ASE: provide a way to label D without changing ASE
Conventionally, ASE's source code needs to be modified to support
recognizing D. This change makes it possible to recognize D without
changing the source code of ASE. The way is to simply modify the mass of
H atoms. If the mass is 2.014 amu, then it is automatically changed to
'H+' for PySCF (which is just D as defined by Xi Xu).

Examples::
>>> from ase.io import read
>>> from pyscf.neo.ase import Pyscf_NEO
>>> atoms = read('example.xyz')
>>> masses = atoms.get_masses()
>>> masses[0] = 2.014 # if the first atom is H and want to change to D
>>> atoms.set_masses(masses)
>>> atoms.calc = Pyscf_NEO(quantum_nuc=['H', 'H+'])