NEO: correct the mass for hydrogen
Previously, the proton mass was used. Note that hydrogen is heavier than
proton because of the additional electron.

It may appear strange to use proton+electron mass for proton's equation,
while this is actually the correct way. It is because the proton's
equation is actually proton+electron center-of-mass equation. Consider
the hydrogen atom, the electronic equation is the relative motion
between the electron and the proton, then the remaining nuclear wave
function is actually fot the COM motion. It is because the proton mass
is significantly larger than the electron mass such that the electron
mass is directly used in the electronic equation (actually it should be
the reduced mass) and the proton mass is approximately correct for the
proton equation. To be precise, here I change the proton mass to
hydrogen mass, which is consistent with the element mass that PySCF
uses.