NEO: use nuclear mass instead of atomic mass
Achieve the consistency in the other way, by removing electron mass from
D atom mass and use proton mass for H (in this case, I just use H atom
mass minus electron mass). Previously I used atom masses for both cases.
Before that, Xi used proton mass and D atom mass.

For full quantum calculations, this is still not ideal, because we are
using the isotope average mass. It seems pointless to me to subtract
electron mass from an average mass instead of pure isotope mass, but
real Hessian analysis and MD simulations all use isotope average masses,
(and they do not remove electron masses) except in some special cases
where certain isotopes are studied.

This will cause some "inconsistency" when comparing with traditional DFT
Hessian analysis, because different masses are used for the mass-scaled
coordinates, not only the underlying PESs are different.

This commit also changes to the behavior that, if H is not quantum
(though impossible for NEO/CNEO calculations), its mass is the isotope
average mass. If H*/H+/H# is requested manually but not set to quantum,
the atom mass is used (with the extra electron mass), though I don't
think this use case is possible.