NEO: A (nearly full) reimplementation of neo moledule
As motivated long ago by a question of @tculpitt "Why do you need to do
so many code modifications just to add positron?", and pushed by the
fact I had to attend the 2nd PySCF conference (such that I can post some
good-looking code), I finally decided to move forward from the original
framework that @xu-xi created and make the neo moledule slightly more
general.

The `mole` submodule, however, is still extremely specific. I think
there is no way to get around this problem. We have to be very specific
to create the multi-component system we want. Further improvements might
be possible but currently I have already tried my best. This part is
mostly similar to the original mole-submole structure, but instead of
using elec/nuc attributes, all components are now origanized in the
`components` dictionary. Dictionary in my opinion is better than list
for this problem as we can easily lose track what is what in a list.

The SCF, gradient and hessian parts are now as general as I can possibly
come up with. They don't see electrons or quantum nuclei (except for
the special treatment that nuclei are singly occupied and has no
self-interactions). They just see general charge/mass particles with
basis functions provided, and calculate intra-component hcore/vhf and
inter-component interactions. Certainly CDFT/gradient/hessian have their
specific position constraints for quantum nuclei, but this specificity
is what we need. In principle with the code now we should be able to
calculate grad/hessian with positrons present, and I hope the code can
even consider grad/hessian with ghost nuclear basis functions, but I
haven't tested these functionalities and they may have some small bugs
to be found and fixed (if someone really does that).

For the other submodules (mostly post-SCF), however, I ran out of time
and I could not improve them on my own. NEO-MP2/solvent/QC are not my
code and help is needed to adapt to the new framework. Solvent test now
has a small error for the energy, but the numerical gradient test now
passes! Maybe the current status is more correct? There is no NEO-MP2
test, but the code in mp2.py runs fine and the result is consistent with
before. QC also runs fine and the tests can pass. My FCI code runs fine
and the result in the N-2 resolution example did not change. Though all
of these rely on the compatibility code that is expected to be removed
in the future.

One TODO for QC: the example originally in hf.py that trims py/pz
orbitals to create a px-only calculation.

In the following I will also summarize some notable changes:

- For the mf.mf_elec.xc = 'XXX' problem, now I have changed the code to
  force the user providing xc= in the mf initialization step. This
  decision is based on the past observations that changing an important
  parameter by changing attributes can lead to possible user errors when
  they completely forget to do so.

- Now `unrestricted` parameter only means we want to force the mf's to
  be unrestricted. It cannot be used to measure if the mf's are
  unrestricted or not, because even if it is False, the mf's can still
  be unrestricted because of its spin. In hf.py there is code to
  automatically decide UHF/RHF cases. Measuring if mf is an instance of
  scf.uhf.UHF is currently my choice.

- I changed PySCF's own CPHF solver due to its issue 2702. Will wait to
  see what is really happening.

- EPC calculations should be faster than before for multi-proton cases.
  I haven't benchmarked it yet.

- Direct Vee calculations for e-n Coulomb should be faster as the new
  code avoids calculation the integrals twice in one SCF cycle.