CNEO-CPHF: change position constraint scale factor
This scale factor is highly empirical, and the 1/(LUMO-HOMO) choice does
work well for the full quantum case (small error, though with a lot of
iterations), but becomes completely nonconvergent for a large molecule
with multiple hydrogen atoms. Now scale down the hydrogen scale factor,
but increase heavy nuclear scale factors. This is likely to fail again
if someone decides to calculate ten full quantum F atoms, but
fortunately no one will actually do that. This is similar to the old
implementation that uses 2*charge as the factor.

Also add more debug output for CPHF calculations. Needs verbose=6 for
hessian object.