NEO: proper ghost atom DFT grid setup
This commit partly finishes the TODO raised in the last commit, by
hacking the upstream code. It creates an electronic mole that appears to
have normal elements, but actually the charge is zero.

This relies on the fact that DFT grid setup checks the symbol then look
up the radius according to the charge associated with the symbol. There
was one piece of code that is consistent with this behavior, which is
the grid response gradient. Now I have changed it to be consistent. This
means that if in the future the upstream decides to directly look up the
atom_charges array instead of the symbols, this hack will fail. Any way,
at least the code in the present status works. The desired feature is
that the ghost atom and the original atom have exactly the same
behavior. See the added example that swapping their coordinates results
in an identical energy. It was not the case before this commit because
ghost atoms were using a larger radius when setting up the meshgrid.

TODO: generalize the upstream DFT grid setup code to really support a
custom radius.