NEO: integrate vint into veff
This is similar to the old design that vint was in hcore, but I think
veff should be a better place to put vint into.

This will inevitably involve the usage of cache, because veff only
accepts the density matrix input of the current component, but vint
requires density matrices of all components. This may lead to desync
between veff and vint, when the density matrix changes but the user
forgot to call get_vint before get_veff to update the cache. This is,
however, 100% the duty of users that they need to figure out if they
need to update vint cache before get_veff calls. Usually it wouldn't
have been necessary because calling get_veff outside the SCF loop is
mostly for post-SCF treatments, and in that case the density matrix
enters get_veff is the converged density matrix, and other components
are considered frozen. This is especially true if epc is not used. In
that case, even if you change the current component's density matrix,
the mean-field Coulomb interactions generated from other components are
precisely the frozen nuclei approximation. When epc is enabled, calling
get_veff alone is certainly a crude approximation.

The reason for making this change is to achieve better compatibility
with upstream PySCF functionalities for the electronic mean-field
object. For example, stability analysis and possibly post-SCF
correlation energy calculation for the electronic component. This is
because there is no way to pass the hcore+veff+vint fock to these
functionalities, and these modules, sometimes uses get_fock, and for
most of the time they will simply call get_hcore + get_veff to construct
the Fock matrix on-the-fly. This was not an issue when vint was in
hcore, but it became an issue when I separated vint out. Now these
should work again because vint is now in veff.