CNEO-FCI: improve c-davidson convergence
Bounded least squares optimization is found to be both robust and
accurate. The selected parameter set can yield position errors close to
machine precision with a reasonable amount of cycles.

Update the output, use logger instead of print.

Update the interface such that it is easy to modify the parameters for
the constrained Davidson iteration.

Co-authored-by: zc62 <17048823+zc62@users.noreply.github.com>