CNEO: fix scanner bug due to unexpected rebound
Previously, using `=` creates a completely new `interactions` dict,
however, components are modified in-place. This creates inconsistencies
when both scanners are being used, and they have shallow copies of these
dictionaries. When rebound occurs in one scanner, the other scanner has
its attributes half correct and half wrong.

Now change to the behavior that everything is mutated in-place.

It is, however, still not recommended to use two scanners at the same
time. The SCF intermediates _eri gets cleared on every reset, hence SCF
must be run multiple times, which is a waste.